feat : ArbreController

+++ Citronix/Citronix/src/main/java/com/app/Citronix/Model/DTO/Response/ArbreResponseLight.java
@@ -9,5 +9,5 @@ import java.time.LocalDate;
 @Builder
 public class ArbreResponseLight {
     private LocalDate datePlantation;
-    private Double age;
+    private long age;
 }
